refactor(transporte): unificar módulo de descripciones y agregar CRUD dinámico
- Se migró la lógica de descripciones al paquete central .
- Se dividió el controlador en  y  aplicando el principio de responsabilidad única.
- Se reemplazaron múltiples funciones GET por funciones inteligentes ( y ) con filtros dinámicos y JOINs condicionales para evitar errores SQL.
- Se completó el CRUD (POST, GET, PUT, DELETE) para los metadatos de líneas y estaciones.
- Se actualizaron los modelos en GORM (, ) fijando llaves foráneas y tipos de datos ().
- Se refactorizaron las rutas (, ) integrándolas de forma limpia con el middleware  y eliminando anidaciones excesivas.

+++ cmd/pkg/models/DescripcionEstacion.go
@@ -10,5 +10,5 @@ type DescripcionEstacion struct {
 	Estado_ciudad      string   `json:"estado_ciudad"`
 	Longitud           float64  `json:"longitud"`
 	Estacion           Estacion `gorm:"foreignKey:EstacionID;references:ID" json:"-"`
-	EstacionID         int16    `json:"estacion_id"`
+	EstacionID         uint     `json:"estacion_id"`
 }
